geotiff: extract input validators to _validation.py (#1882)
Step 4 of #1813's multi-PR refactor of __init__.py. Pure code motion;
no public API change.

Moved into a new xrspatial/geotiff/_validation.py:

- _validate_3d_writer_dims: rejects ambiguous (band,y,x) vs (y,x,band)
  on 3D writer input (#1812).
- _validate_dtype_cast: blocks lossy float-to-int casts on read.
- _validate_tile_size: positive int + multiple-of-16 for tiled writers
  (TIFF 6 TileWidth/TileLength requirement, #1767).
- _validate_chunks_arg: positive int / 2-tuple of positive ints for
  every dask read entry point (#1752, #1776).
- _validate_tile_size_arg: backwards-compat wrapper for
  _validate_tile_size.

All five are pure leaves over numpy dtypes and Python primitives;
only dependencies are np, _BAND_DIM_NAMES from _coords, and
_X_DIM_NAMES/_Y_DIM_NAMES from _runtime (#1880). __init__.py
re-imports each name so every existing caller resolves unchanged.

Net __init__.py change: 4726 -> 4569 lines (-157). _validation.py is
181 lines.

Verification: pixel-parity matrix from #1889 (192 cells), runtime
identity tests from #1890 (10), and the validator-specific tests
(test_size_param_validation_1752.py,
test_size_param_validation_gpu_vrt_1776.py,
test_tile_size_multiple_of_16_1767.py,
test_to_geotiff_3d_dim_validation_1812.py,
test_geotiff_band_bool_rejection_1786.py,
test_geotiff_planar_strip_truncation_1782.py) all pass (302/302
green, 1 pre-existing main failure unrelated to this PR).

Refs #1813.
